requirements-dev.txt
coverage>=7,<8
jsonschema>=4.10,<5
PyYAML>=6,<7
